Tried to fix rounding error

+++ geom.cpp
@@ -36,7 +36,7 @@ int collinear(point2d p, point2d q, point2d r) {
 /* return 1 if c is  strictly left of ab; 0 otherwise */
 // Signed Area should be positive if c is to the left
 int left_strictly(point2d a, point2d b, point2d c) {
-  int signed_area = signed_area2D(a, b, c);
+  double signed_area = signed_area2D(a, b, c);
   if (signed_area > 0){
     return 1;
   }
@@ -46,7 +46,7 @@ int left_strictly(point2d a, point2d b, point2d c) {
 
 /* return 1 if c is left of ab or on ab; 0 otherwise */
 int left_on(point2d a, point2d b, point2d c) {
-  int signed_area = signed_area2D(a, b, c);
+  double signed_area = signed_area2D(a, b, c);
   if (signed_area > 0 || signed_area == 0){
     return 1;
   }
